+++ src/MicroElements.Functional/EnumerableExtensions.cs
@@ -4,6 +4,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using JetBrains.Annotations;
 
 namespace MicroElements.Functional
 {
@@ -18,7 +19,8 @@ namespace MicroElements.Functional
         /// <typeparam name="T">Item type.</typeparam>
         /// <param name="items"><see cref="IEnumerable{T}"/> or null.</param>
         /// <returns>The same items or empty enumeration.</returns>
-        public static IEnumerable<T> NotNull<T>(this IEnumerable<T>? items) =>
+        [LinqTunnel]
+        public static IEnumerable<T> NotNull<T>([NoEnumeration] this IEnumerable<T>? items) =>
             items ?? Array.Empty<T>();
 
         /// <summary>