+++ src/MicroElements.Functional/SharedUtils/TypeExtensions.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
+using System.Diagnostics.CodeAnalysis;
 using System.Reflection;
 
 namespace MicroElements.Functional
@@ -16,8 +17,9 @@ namespace MicroElements.Functional
         /// </summary>
         /// <param name="type">Source type.</param>
         /// <returns>Default value.</returns>
-        public static object? GetDefaultValue(this Type type)
+        public static object? GetDefaultValue([DisallowNull] this Type type)
         {
+            type.AssertArgumentNotNull(nameof(type));
             return type.IsValueType ? Activator.CreateInstance(type) : null;
         }
 
@@ -27,10 +29,9 @@ namespace MicroElements.Functional
         /// <typeparam name="T">The type to test assignability to.</typeparam>
         /// <param name="type">The type to check.</param>
         /// <returns>True if type is assignable to references of type <typeparamref name="T" />; otherwise, False.</returns>
-        public static bool IsAssignableTo<T>(this Type type)
+        public static bool IsAssignableTo<T>([DisallowNull] this Type type)
         {
             type.AssertArgumentNotNull(nameof(type));
-
             return typeof(T).GetTypeInfo().IsAssignableFrom(type.GetTypeInfo());
         }
 
@@ -39,7 +40,7 @@ namespace MicroElements.Functional
         /// </summary>
         /// <param name="type">The type to check.</param>
         /// <returns>True if type is concrete.</returns>
-        public static bool IsConcreteType(this Type type)
+        public static bool IsConcreteType([DisallowNull] this Type type)
         {
             type.AssertArgumentNotNull(nameof(type));
             return !type.IsInterface && !type.IsAbstract;